Toggles for Paid Checkboxes and hide non-respondents
Done in conjunction with update to picks.gs update allowing members to toggle visibility of non-respondent rows on sheet import (only in configuration settings, not real-time). Paid column as an optional import feature was implemented in the weekly sheet generation and turning it on is now an option on the configuration.