Updated Score Fetch
Issue where the order of API outcomes for the endpoint were not in chronological order for some members (maybe time zone related?). Now the script checks for the column index of the final matchup, looks for that within the "completedGames" array of objects, then uses the values from that object if present to set the tiebreaker.